Check to make sure that the fallback lang wasn't returned when saving new ones
Because the fallback locale is now returned when a requested one isn't found
when an attempt to save a new locale is made, it would overwrite the fallback.
This means that we need to check that the returned langModel is actually the
same language of the one we are trying to save.